Give warm claim and completion SQL one owner #163
